requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
psutil
bs4
lxml
psycopg2-binary
py-cpuinfo
dblpy
